fixed issues with multiple budget setup
