+++ client/index.js
@@ -1,25 +1,26 @@
 // Screens
-export { default as Home } from "./screens/Home";
-export { default as Navbar } from "./screens/Navbar";
+export { default as Home } from './screens/Home';
+export { default as Navbar } from './screens/Navbar';
 
 // Accounts
-export { default as Accounts } from "./screens/Accounts";
-export { default as Individual } from "./screens/Accounts/Individual";
-export { default as IndividualAccount } from "./screens/Accounts/IndividualAccount";
+export { default as Accounts } from './screens/Accounts';
+export { default as Individual } from './screens/Accounts/Individual';
+export {
+  default as IndividualAccount
+} from './screens/Accounts/IndividualAccount';
 
 // Budget
-export { default as Budget } from "./screens/Budget";
-export { default as BudgetSetup } from "./screens/Budget/Setup";
-export { default as EditCategories } from "./screens/Budget/EditCategories";
+export { default as Budget } from './screens/Budget';
+export { default as BudgetSetup } from './screens/Budget/BudgetSetup';
+export { default as EditCategories } from './screens/Budget/EditCategories';
 
 //OnBoard
-export { default as Initial } from "./screens/OnBoard/Initial";
-export { default as Link } from "./screens/OnBoard/Link";
-export { default as Login } from "./screens/OnBoard/Login";
-export { default as Quiz } from "./screens/OnBoard/Quiz";
-export { default as Result } from "./screens/OnBoard/Result";
-export { default as Signup } from "./screens/OnBoard/Signup";
+export { default as Initial } from './screens/OnBoard/Initial';
+export { default as Link } from './screens/OnBoard/Link';
+export { default as Login } from './screens/OnBoard/Login';
+export { default as Quiz } from './screens/Home/Quiz';
+export { default as Result } from './screens/Home/Result';
+export { default as Signup } from './screens/OnBoard/Signup';
 
 // Profile
-export { default as Profile } from "./screens/Profile";
-
+export { default as Profile } from './screens/Profile';
