+++ client/screens/Budget/BudgetSetup.js
@@ -1,9 +1,9 @@
 import React from 'react';
-import { View, Text, Image } from 'react-native';
-import { FormLabel, FormInput, Button } from 'react-native-elements';
+import { View, Text } from 'react-native';
+import { FormInput, Button } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { setBudget } from '../../store';
-import { styles } from '../../common/styles';
+import { styles, colorTheme } from '../../common/styles';
 
 class BudgetSetup extends React.Component {
   constructor() {
@@ -12,46 +12,56 @@ class BudgetSetup extends React.Component {
       income: 0,
       staticCosts: 0,
       savings: 0,
-      spendingBudget: 0,
+      spendingBudget: 0
     };
   }
+  static navigationOptions = {
+    headerStyle: { backgroundColor: colorTheme.blue.medium }
+  };
 
   render() {
     return (
       <View style={styles.container}>
-        <View style={styles.logoLocation}>
-          <Image source={require('../../../public/img/logo.png')} />
-          <Text style={styles.h1}>
-            Tell us a little about your finances.
-          </Text>
-        </View>
+        <Text style={styles.initialScreenText}>
+          Tell us a little about your finances.
+        </Text>
         <Text style={styles.budgetSetupText}>What is your income?</Text>
         <FormInput
-          containerStyle={{ width: '80%', paddingTop: 5 }}
+          containerStyle={{ width: '80%', paddingTop: 2 }}
           inputStyle={styles.formInput}
           onChangeText={income => this.setState({ income: +income })}
-          value={this.state.income}
           placeholder="Income"
         />
         <Text style={styles.budgetSetupText}>
           What are your static costs? (i.e. rent, utilities, insurance, etc.)
         </Text>
         <FormInput
-          containerStyle={{ width: '80%', paddingTop: 10 }}
+          containerStyle={{ width: '80%', paddingTop: 2 }}
           inputStyle={styles.formInput}
           onChangeText={staticCosts =>
             this.setState({ staticCosts: +staticCosts })
           }
-          value={this.state.staticCosts}
           placeholder="Static Costs"
         />
-        <Text style={styles.budgetSetupText}>How much would you like to save?</Text>
+        <Text style={styles.budgetSetupText}>
+          How much would you like to save?
+        </Text>
         <FormInput
-          containerStyle={{ width: '80%', paddingTop:5 }}
+          containerStyle={{ width: '80%', paddingTop: 2 }}
           inputStyle={styles.formInput}
           onChangeText={savings => this.setState({ savings: +savings })}
-          value={this.state.savings}
           placeholder="Savings"
+        />
+        <Text style={styles.budgetSetupText}>
+          How much is currently in your retirement savings?
+        </Text>
+        <FormInput
+          containerStyle={{ width: '80%', paddingTop: 2 }}
+          inputStyle={styles.formInput}
+          onChangeText={retirement =>
+            this.setState({ retirement: +retirement })
+          }
+          placeholder="Retirement"
         />
         <View style={{ padding: 10 }}>
           <Button
@@ -64,12 +74,10 @@ class BudgetSetup extends React.Component {
                 this.state.income - this.state.staticCosts - this.state.savings;
               this.props.setBudget({ ...this.state, spendingBudget });
               this.props.navigation.navigate('EditCategories', {
-                title: 'EditCategories',
+                title: 'EditCategories'
               });
             }}
-          >
-            Submit
-          </Button>
+          />
         </View>
       </View>
     );
@@ -78,13 +86,13 @@ class BudgetSetup extends React.Component {
 
 const mapState = state => {
   return {
-    budget: state.budget,
+    budget: state.budget
   };
 };
 
 const mapDispatch = dispatch => {
   return {
-    setBudget: budget => dispatch(setBudget(budget)),
+    setBudget: budget => dispatch(setBudget(budget))
   };
 };
